+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 - {resume|suspend}.py ignore nodes not in cloud configuration (config.yaml).
 - Constrain packages in Pipfile and requirements.txt
 - Allow hybrid scripts to succeed when slurm user does not exist.
+- Fix pushing cluster config to project metadata in hybrid terraform
+  deployments.
 
 ## \[5.0.2\]